correção de crud no angular

+++ gestao-service/src/main/java/com/natan/gestaoservice/repository/PessoaRepository.java
@@ -21,6 +21,6 @@ public interface PessoaRepository extends Repository<Pessoa, Long> {
 	
 	List<Pessoa> findAll();
 
-	Pessoa findById(Long codigo);
+	Pessoa findById(Integer codigo);
 
 }
